refactor(phases): migrate to PhaseAsset ScriptableObjects
- Convert concrete phase classes (e.g., GainPhase, ActionPhase) into ScriptableObjects (PhaseAsset) to allow editor configuration.
- Move phase-specific event listener registration from constructor to Run() method to ensure proper initialization at runtime.
- Abstract Type property in PhaseAsset for each phase to define its PhaseType in code, replacing serialized type field.
- Improves safety, flexibility, and clarity of phase lifecycle management while keeping IPhase interface intact.
