+++ Vermines/Assets/Resources/ScriptableObject/Game/Phases.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37a1d7b38a646ad4ab715e25b43e0321
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
