+++ Vermines/Assets/Resources/ScriptableObject/Game/Phases/ActionPhaseAsset.asset
@@ -1,0 +1,15 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a4de7e66d532f344788de5678493cd6d, type: 3}
+  m_Name: ActionPhaseAsset
+  m_EditorClassIdentifier: 
+  type: 2
