+++ Vermines/Assets/Resources/ScriptableObject/Game/Phases/ActionPhaseAsset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9236617600cc4234a98c920257cc2fcb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
