+++ Vermines/Assets/Resources/ScriptableObject/Game/Phases/GainPhaseAsset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc7aa5203e724e14d90cea4e8d89882e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
