+++ Vermines/Assets/Resources/ScriptableObject/Game/Phases/ResolutionPhaseAsset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 312b28fe26f967b4dbd4dafc3c64c2f8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
