+++ Vermines/Assets/Resources/ScriptableObject/Game/Phases/SacrificePhaseAsset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e3ef322ecf154941a938e92a619d672
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
