+++ Vermines/Assets/Scripts/Gameplay/Cards/Effects/ShopEffect/ReplaceShopEffect.cs
@@ -53,6 +53,8 @@ namespace Vermines.Gameplay.Cards.Effect {
 
         private void ReplaceCard(Dictionary<ShopType, int> dictShopSlot)
         {
+            if (UIContextManager.Instance)
+                UIContextManager.Instance.PopContextOfType<ReplaceEffectContext>();
             // Check if the dictionary is null or empty
             if (dictShopSlot == null || dictShopSlot.Count == 0) {
                 Debug.LogWarning("[ReplaceEffect] dictShopSlot is null or empty, nothing to replace.");
