--- Vermines/Assets/Scripts/Gameplay/Logic/Phase/APhaseBase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8cb56974de765d04d9d11a5f31c9896d
