--- Vermines/Assets/Scripts/Gameplay/Logic/Phase/ActionPhase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e98c9bde63c6c4d4a9b112933696efb5
