--- Vermines/Assets/Scripts/Gameplay/Logic/Phase/GainPhase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 158a4b854a31ea44c9d4247b5927574a
