--- Vermines/Assets/Scripts/Gameplay/Logic/Phase/ResolutionPhase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: afdc4f082f4554b488f2c95d22940e26
