--- Vermines/Assets/Scripts/Gameplay/Logic/Phase/SacrificePhase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5b3f5e5c1aabeba4896107aa5f648c0c
