+++ Vermines/Assets/Scripts/UI/HUD/GameplayUIController.cs
@@ -1,11 +1,12 @@
 ﻿using Fusion;
-using System.Collections.Generic;
-using UnityEngine;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using UnityEngine;
 
 namespace Vermines.UI
 {
+    using System.Collections;
     using Vermines.UI.Screen;
 
     public class GameplayUIController : MonoBehaviour
@@ -209,7 +210,6 @@ namespace Vermines.UI
         {
             if (_ActiveScreen != null)
             {
-                Debug.Log($"Hiding screen: {_ActiveScreen.GetType().Name}");
                 _ActiveScreen.Hide();
                 _ActiveScreen = null;
             }
@@ -338,6 +338,27 @@ namespace Vermines.UI
             }
 
             return _DualPopupHandler.OpenDualButtonPopupAsync(message, header, leftButtonLabel, rightButtonLabel);
+        }
+
+        public static void StartTurnUIRoutine()
+        {
+            GameplayUIController controller = GameObject.FindAnyObjectByType<GameplayUIController>(FindObjectsInactive.Include);
+            if (controller != null)
+                controller.StartCoroutine(controller.SacrificeRoutine());
+        }
+
+        private IEnumerator SacrificeRoutine()
+        {
+            GameplayUIController gameplayUIController = GameObject.FindAnyObjectByType<GameplayUIController>(FindObjectsInactive.Include);
+
+            if (gameplayUIController != null)
+                gameplayUIController.Show<GameplayUITurn>();
+
+            yield return new WaitForSeconds(3f);
+
+            if (gameplayUIController != null && gameplayUIController.GetActiveScreen(out GameplayUIScreen activescreen) &&
+                activescreen is GameplayUITurn)
+                gameplayUIController.Hide();
         }
 
         #endregion
